implemented uint8,16,32,64, complex32,63,128...we can now only create…
tensors, but operations won't support, will make this too once
completing f8,f4

+++ Tests/ReductionsTests/stat_test.cpp
@@ -9,7 +9,7 @@ int main()
     opts.device = Device::CPU;
     opts.requires_grad = true;
 
-    Tensor A(Shape{{3,3}}, Dtype::Int32, DeviceIndex(Device::CPU));
+    Tensor A(Shape{{3,3}}, Dtype::Float32, DeviceIndex(Device::CPU));
     //std::vector<int> dataa= {1,2,3,4,5,6};
     //A.set_data(dataa);
     A.fill(12132);
@@ -48,7 +48,7 @@ A.display(std::cout, 5);
     opts.device = Device::CUDA;
     opts.requires_grad = true;
 
-    Tensor A1(Shape{{3,3}}, Dtype::Int32, DeviceIndex(Device::CUDA));
+    Tensor A1(Shape{{3,3}}, Dtype::Float32, DeviceIndex(Device::CUDA));
     //std::vector<int> data = {1,2,3,4,5,6};
     //A1.set_data(data);
     A1.fill(12132);
